% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nls_formula.R
\name{build_nls_formula}
\alias{build_nls_formula}
\title{Helper function to build nls spatial formulas}
\usage{
build_nls_formula(lags, covariates, y_name, model = NULL)
}
\description{
Helper function to build nls spatial formulas
}
