% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts}
\alias{impacts}
\title{Estimate spatial and spatiotemporal effects}
\usage{
impacts(model, variable = NULL, W, lag_dv = NULL, groups = NULL)
}
\arguments{
\item{model}{Model object estimated with \code{lagsarnls}}

\item{variable}{Vector of one or more characters matching variable names in model
formula. If NULL (default), the function returns impact estimates for all
variables.}

\item{W}{Spatial weights matrix.}

\item{lag_dv}{Character string naming lagged dependent variable in model.}

\item{groups}{Vector of IDs for cross-sectional units in data. NOTE: the
number of unique groups must match the dimensions of W
(i.e, nrow(W) == length(unique(groups))), or the function will return
an error. In this case, subset your W matrix for ONE temporal
cross-section.}
}
\value{
Returnd a data frame with estimated impacts
}
\description{
\code{impacts} estimates dynamic marginal effects for spatial and
spatiotemporal dynamic regression models estimated with \code{\link{lagsarnls}}.
For spatial models (i.e., no temporal lag), the function returns the spatial
impacts (direct, indirect, and total) for unit-changes in exogenous variables.
For spatiotemporal models (i.e., those with temporal lags), the function returns
the long-run steady state (LRSS) response to unit-changes in exogenous variables.
}
\examples{
 # Simulate spatiotemporal data using defaults from \code{spacetime_sim2}
 x = spacetime_sim2()

 # Estimate model
 m = lagsarnls(formula = y ~ ylag + x1 + x2,
               data    = x$data,
               W       = x$W_matrix)

 # Subset w for ONE cross-section
 n  = length(unique(x$data$Group))
 W2 = x$W_matrix[1:n,1:n]

 impacts(model  = m,
         W      = W2,
         lag_dv = "ylag",
         groups = x$data$Groups)

}
