% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts_ci.R
\name{impacts_ci}
\alias{impacts_ci}
\title{Estimate credibility intervals for spatial and spatiotemporal impacts using
parametric simulation.}
\usage{
impacts_ci(
  model,
  W,
  beta,
  phi = NULL,
  rho = "rho",
  unit_ids = NULL,
  time_ids = NULL,
  parametric = TRUE,
  truncated_normal = TRUE,
  cred_interval = 0.95,
  sims = 100,
  seed = 1
)
}
\arguments{
\item{model}{Model object of class nls estimated with
\code{\link{lagsarnls}}}

\item{W}{Spatial weights matrix SUBSET for one cross-section if
spatiotemporal}

\item{beta}{Character string indicating variable to estimate impacts}

\item{phi}{Character string indicating name of temporally lagged dependent
variable}

\item{rho}{Character string indicating name of spatial lag (set to "rho" by
default)}

\item{unit_ids}{Vector of unit IDs from panel data}

\item{time_ids}{Vector of time IDs from panel data}

\item{parametric}{Not used. Relevant for delta method implementation in
the future}

\item{truncated_normal}{Boolean. Use truncated normal for sampling? Default
True. See details for more.}

\item{cred_interval}{Credibility interval to estimate as a decimal
(0.95 ~ 95\% by default)}

\item{sims}{Simulation iterations}

\item{seed}{Set seed (relevant for replication purposes)}
}
\value{
Returns a data frame of Direct, Indirect, and Total Impact estimates
along with simulated credibility intervals. "Estimate" reflects the median
response from simulations while the Lower and Uppwer bounds are the quantile
at probabilities specified by the user (95\% default). Quantiles help to
further mitigate the problems of explosive parameter draws.
}
\description{
This function uses a parametric simulation to estimate
credibility intervals for impacts from a spatial or spatiotemporal
regression model estimate with \code{\link{lagsarnls}}.
}
\details{
Note: Sampling from a truncated normal distribution
(\code{truncated_normal = TRUE}, i.e., the default) is highly recommended.
Sampling from a standard multivariate normal may result in sampled values of
Rho or Phi which exceed the theoretical bounds of \link{-1,1} resulting in an
explosive system and invalid quantites of interest. To avoid this possibility,
the truncated normal sampling option limits draws for Rho and Phi between
-1 and 1.
}
\examples{
m <- lagsarnls(formula = y ~ ylag + x1 + x2,
               data    = x$data,
               W       = x$W_matrix)
n  <- length(unique(x$data$Group))
W2 <- x$W_matrix[1:n,1:n]

effs <- impacts_ci(model            = m,
                   W                = W2,
                   beta             = "x1",
                   phi              = "ylag",
                   rho              = "rho",
                   unit_ids         = x$data$Group,
                   time_ids         = x$data$Time,
                   parametric       = TRUE,
                   truncated_normal = TRUE,
                   cred_interval    = 0.95,
                   sims             = 100,
                   seed             = 123)

print(effs)
}
