% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagsarnls.R
\name{lagsarnls}
\alias{lagsarnls}
\title{Spatial lag models using non-linear least squares}
\usage{
lagsarnls(
  formula = NULL,
  W = NULL,
  data = NULL,
  subset_W = FALSE,
  lags = 6,
  intercept = TRUE,
  start_vals = NULL,
  rho_start = 0.2,
  rho_bounds = NULL
)
}
\arguments{
\item{formula}{model formula}

\item{W}{spatial weights matrix}

\item{data}{data}

\item{subset_W}{not implemented}

\item{lags}{spatial lags to approximate}

\item{intercept}{boolean - to include intercept}

\item{start_vals}{starting values for nls optimizer. if null uses ols estimates}

\item{rho_start}{starting values for rho}

\item{rho_bounds}{lower and upper bounds for rho (not necessary)}
}
\value{
returns a nls model object
}
\description{
This function provides functionality for spatial lag models
estimated at high speed using nonlinear least squares.
}
\examples{
x = spacetime_sim()
m = lagsarnls(formula = y ~ x1 + x2,
              data    = x$data,
              W       = x$W_matrix)

}
