% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_w.R
\name{simulate_W}
\alias{simulate_W}
\title{Helper function to generate W in spacetime_sim}
\usage{
simulate_W(n, nneigh, seed = 123)
}
\arguments{
\item{n}{number of observations in W}

\item{nneigh}{number of nearerst neighbors}

\item{seed}{simulation seed for setting random draws}
}
\value{
Returns sparse weights matrix
}
\description{
Not exported. Function borrowed (directly) from Davide Martinetti
and Ghislain Geniaux's package "ProbitSpatial" which is no longer (2021-02-23)
hosted on CRAN. All code credit belongs to them.
}
