% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_sim.R
\name{spacetime_sim}
\alias{spacetime_sim}
\title{Simulate spatial or spatiotemporal data}
\usage{
spacetime_sim(
  sample_size = 100,
  k = 2,
  b_true = c(1, 2.5, -1),
  rho_true = 0.6,
  phi_true = 0.3,
  prob_connect = 0.3,
  e_sd = 0.2,
  probit = FALSE,
  temporal = FALSE,
  groups = NULL,
  time_periods = NULL
)
}
\arguments{
\item{sample_size}{sample size}

\item{k}{number of vars}

\item{b_true}{true beta values}

\item{rho_true}{true rho (spatial) correlation}

\item{phi_true}{true phi (temporal) correlation (not working)}

\item{prob_connect}{density of connections in W}

\item{e_sd}{noise in the dgp}

\item{probit}{boolean simulate binary outcome data}

\item{temporal}{boolean simulate temporal data}

\item{groups}{number of groups if temporal is true}

\item{time_periods}{number of time periods if temporal is true}
}
\value{
A list consisting of 4 elements: data as a tibble, simulated
parameters (user set and observed values), w, w as a list object
}
\description{
spacetime_sim - easy simulation of spatial or spatiotemporal data.
}
