% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scacetime_sim2.R
\name{spacetime_sim2}
\alias{spacetime_sim2}
\title{Simulate spatiotemporal data}
\usage{
spacetime_sim2(
  rho_true = 0.3,
  phi_true = 0.8,
  b_true = c(1, 2.5, -1),
  k = 2,
  e_sd = 1,
  intercept = TRUE,
  panel_units = 20,
  panel_times = 10,
  forfeit_lag = FALSE,
  unit_effects = T,
  time_effects = T,
  unit_betas = NULL,
  time_betas = NULL,
  prob_connect = 0.3
)
}
\arguments{
\item{rho_true}{Measure of spatial dependence (-1 to 1)}

\item{phi_true}{Measure of temporal dependence (-1 to 1)}

\item{b_true}{True paramaeter values for intercept (if included) and Xs (must
equal length of k+1)}

\item{k}{Number of exogenous independent variables to include}

\item{e_sd}{Disturbance term standard deviation (assumed homoskedastic)}

\item{intercept}{Boolean to include intercept}

\item{panel_units}{Number of units in panel}

\item{panel_times}{Number of time periods in panel}

\item{forfeit_lag}{Boolean to forfeit temporal lag. FALSE by default. Mostly
to experiment with consequences of dropping first lagged unit on model
performance.}

\item{unit_effects}{Boolean to include unit-fixed-effects}

\item{time_effects}{Boolean to include time-fixed-effects}

\item{unit_betas, time_betas}{Vector for unit or time FE betas if included in DGP. Must equal length
(unit_effects - 1) or (time_effects - 1) to avoid perfect collinearity.
If NULL (default), random values are generated and returned.}

\item{prob_connect}{Percentage of connections in matrix W, 0.3 by default.
Higher values can help to avoid "island" cases in W where a unit has no
neighbors.}
}
\value{
Returns a list containing a dataframe with simulated spatiotemporal
data, input parameters, and spatial weights information in matrix and
spdep list formats.
}
\description{
spacetime_sim2 simulates spatiotemporal data with the option of
unit and time effects in addtion to space and time dynamic processes.
}
\examples{
# Rapid simulation using defualt values
x = spacetime_sim2()
head(x$data)

}
