% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_create.R
\name{w_create}
\alias{w_create}
\title{w create}
\usage{
w_create(
  formula,
  data,
  sp_data,
  unit_id = NULL,
  time_id = NULL,
  style = "W",
  nb_list = FALSE,
  silent = FALSE
)
}
\arguments{
\item{formula}{A formula class object which will be used for subsetting data}

\item{data}{The data frame containing rows which will be subset}

\item{sp_data}{Simple feature object containing geometry to identify neighbors}

\item{unit_id}{A character string identifying units in the data}

\item{time_id}{A character string identifying times in the data. If supplied,
the function returns a panel spatial weights matrix. Note, for now only
balanced panels are supported - i.e., units must be identical from year-
to-year.}

\item{style}{Definition of neighbors passed to poly2nb() from spedep. Default
is "W" for row-standardized spatial weights.}

\item{nb_list}{Boolean True or False (default) indicating whether the function
should return a spatial weights list object in addition to a spatial weights
matrix}

\item{silent}{Boolean True or False (default) indicating whether the function
should print progress messages to the console.}
}
\value{
Returns a sparse spatial weights matrix and, if requested, a
neighbor list which can be used with models in spatialreg
}
\description{
w_create automates spatial weights construction based on a
supplied model formula. The user supplies a formula class object, their
data, and a simple features object containing the geomety they with to
use for identifying neighbors. The function will return a cross-section
or, if a unique time_id is supplied, a panel spatial weights object.

NOTE - at present, the function will not work wtih unbalanced panels.
}
